Flattern directory structure and move taskrunner to helpers
