+++ nproto/binlogmsg/binlogmsg_test.go
