+++ nproto/binlogmsg/doc.go
