+++ nproto/binlogmsg/entry.go
