+++ nproto/binlogmsg/options.go
