+++ nproto/dbpipe/doc.go
