+++ nproto/dbpipe/msgnode.go
