+++ nproto/durconn/doc.go
