+++ nproto/msgenc/enc.go
@@ -1,4 +1,4 @@
-package enc
+package msgenc
 
 import (
 	"github.com/huangjunwen/nproto/nproto"
